Split SchemaPropertyFunctionalTest to satisfy detekt
The class reached detekt's TooManyFunctions threshold on the
functionalTest source set. Move the DEFINITION_FOR_MAIN_SCHEMA cases
into their own test class.
